2019/requirements.txt
rpi.gpio
python-vlc